Better naming classes with `PDO` - part 2.
