+++ src/Connection.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Oracle;
 
 use Throwable;
-use Yiisoft\Db\Driver\PDO\AbstractConnectionPDO;
-use Yiisoft\Db\Driver\PDO\CommandPDOInterface;
+use Yiisoft\Db\Driver\Pdo\CommandInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidCallException;
@@ -21,9 +20,9 @@ use Yiisoft\Db\Transaction\TransactionInterface;
  *
  * @link https://www.php.net/manual/en/ref.pdo-oci.php
  */
-final class Connection extends AbstractConnectionPDO
+final class Connection extends \Yiisoft\Db\Driver\Pdo\AbstractConnection
 {
-    public function createCommand(string $sql = null, array $params = []): CommandPDOInterface
+    public function createCommand(string $sql = null, array $params = []): CommandInterface
     {
         $command = new Command($this);
