+++ src/Driver.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Oracle;
 
 use PDO;
-use Yiisoft\Db\Driver\PDO\AbstractPDODriver;
 
 /**
  * Implements the Oracle Server driver based on the PDO (PHP Data Objects) extension.
  *
  * @see https://www.php.net/manual/en/ref.pdo-oci.php
  */
-final class Driver extends AbstractPDODriver
+final class Driver extends \Yiisoft\Db\Driver\Pdo\AbstractDriver
 {
     public function createConnection(): PDO
     {
