+++ tests/ConnectionTest.php
@@ -7,7 +7,6 @@ namespace Yiisoft\Db\Oracle\Tests;
 use PDO;
 use Throwable;
 use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Driver\PDO\ConnectionPDOInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
@@ -125,7 +124,7 @@ final class ConnectionTest extends CommonConnectionTest
         $this->assertNotNull($connection->getPDO());
 
         $unserialized = unserialize($serialized);
-        $this->assertInstanceOf(ConnectionPDOInterface::class, $unserialized);
+        $this->assertInstanceOf(ConnectionInterface::class, $unserialized);
         $this->assertNull($unserialized->getPDO());
         $this->assertEquals(123, $unserialized->createCommand('SELECT 123 FROM DUAL')->queryScalar());
         $this->assertNotNull($connection->getPDO());
