+++ tests/SchemaTest.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Oracle\Tests;
 
 use Yiisoft\Db\Command\CommandInterface;
-use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Driver\PDO\ConnectionPDOInterface;
+use Yiisoft\Db\Driver\Pdo\ConnectionInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
@@ -238,7 +237,7 @@ final class SchemaTest extends CommonSchemaTest
 
         $commandMock = $this->createMock(CommandInterface::class);
         $commandMock->method('queryAll')->willReturn([]);
-        $mockDb = $this->createMock(ConnectionPDOInterface::class);
+        $mockDb = $this->createMock(ConnectionInterface::class);
         $mockDb->method('getQuoter')->willReturn($db->getQuoter());
         $mockDb
             ->method('createCommand')
@@ -283,16 +282,5 @@ final class SchemaTest extends CommonSchemaTest
         ]);
         */
         return parent::withIndexDataProvider();
-    }
-
-    public function testNotConnectionPDO(): void
-    {
-        $db = $this->createMock(ConnectionInterface::class);
-        $schema = new Schema($db, DbHelper::getSchemaCache(), 'system');
-
-        $this->expectException(NotSupportedException::class);
-        $this->expectExceptionMessage('Only PDO connections are supported.');
-
-        $schema->refreshTableSchema('customer');
     }
 }
